Refactors with Qt6 - Moved Widgets to Widget - Moved MainWindow to GUI - Renamed AcquisitionEngine to ConfigController - Isolated core parameters to mdacore
