+++ pylab/__main__.py
@@ -2,9 +2,12 @@
 
 #from pylab.cli import main  # pragma: no cover
 import click
-from pylab.utils import utils
-from pylab import mdacore
-
+#from pylab.mdacore import *
+from PyQt6.QtWidgets import QApplication
+from pylab.engine import *
+from pylab.gui import MainWindow, run_gui
+from pylab.config import Config
+from pylab.mdacore import load_dev_cores
 '''
 This is the client terminal command line interface
 
@@ -17,11 +20,34 @@ def cli():
     pass
 
 @cli.command()
-def launch():
+@click.option('--dev', is_flag=True, help='Run in development mode.')
+def launch(dev):
     """
     Launch napari with mesofield acquisition interface widgets
     """
-    mdacore.load_custom_gui()
+    if dev:
+        load_dhyana_mmc_params(mmcore_dhyana)
+        load_thorcam_mmc_params(mmcore_thor)
+        mmcore_dhyana.mda.set_engine(MesoEngine(mmcore_dhyana, use_hardware_sequencing=True))
+        mmcore_thor.mda.set_engine(PupilEngine(mmcore_thor, use_hardware_sequencing=True))
+        mmcore_dhyana.run_mda(useq.MDASequence(time_plan={"interval": 0, "loops": 20000}))#, output=r'C:/dev/dh.ome.tif')
+        mmcore_thor.run_mda(useq.MDASequence(time_plan={"interval": 0, "loops": 20000}))#, output=r'C:/dev/thor.ome.tif')
+    else:
+        run_gui(mmcore_dhyana, mmcore_thor, Config)
+
+@cli.command()
+def dev():
+    """
+    Start the application.
+    """
+    app = QApplication([])
+    core1, core2 = load_dev_cores()
+    cfg = Config
+    mesofield = MainWindow(core1, core2, cfg)
+    mesofield.show()
+    app.exec_()
+
+
 
 ### Utility commands for querying serial ports and USB IDs ###
