+++ pylab/widgets.py
@@ -1,22 +1,23 @@
-from typing import TYPE_CHECKING
-if TYPE_CHECKING:
-    import napari
-
 import pymmcore_plus
-from magicgui import magicgui
-from itertools import count
-
-from pathlib import Path
-import numpy as np
 from pymmcore_plus import CMMCorePlus
-from qtpy.QtWidgets import (
+import os
+import subprocess #for PsychoPy Subprocess
+from magicgui import magicgui
+
+from PyQt6.QtWidgets import (
     QGroupBox,
     QHBoxLayout,
     QLabel,
     QVBoxLayout,
     QWidget,
+    QLineEdit,
+    QPushButton,
+    QComboBox,
+    QTableWidget,
+    QHeaderView,
+    QFileDialog,
+    QTableWidgetItem
 )
-from useq import MDAEvent
 
 from pymmcore_widgets import (
     MDAWidget,
@@ -26,19 +27,24 @@ from pymmcore_widgets import (
     SnapButton,
 )
 
-import useq
-
 from pylab.config import ExperimentConfig
-from pylab.engine import AcquisitionEngine
 
-# Import necessary modules for the IPython console
-from qtconsole.rich_jupyter_widget import RichJupyterWidget
-from qtconsole.inprocess import QtInProcessKernelManager
-from qtpy.QtWidgets import QMainWindow
-import pandas as pd
-import datetime
+@magicgui(call_button='Start LED', mmc={'bind': pymmcore_plus.CMMCorePlus.instance()})   
+def load_arduino_led(mmc):
+    """ Load Arduino-Switch device with a sequence pattern and start the sequence """
+    
+    mmc.getPropertyObject('Arduino-Switch', 'State').loadSequence(['4', '4', '2', '2'])
+    mmc.getPropertyObject('Arduino-Switch', 'State').setValue(4) # seems essential to initiate serial communication
+    mmc.getPropertyObject('Arduino-Switch', 'State').startSequence()
 
-import json
+    print('Arduino loaded')
+
+@magicgui(call_button='Stop LED', mmc={'bind': pymmcore_plus.CMMCorePlus.instance()})
+def stop_led(mmc):
+    """ Stop the Arduino-Switch LED sequence """
+    
+    mmc.getPropertyObject('Arduino-Switch', 'State').stopSequence()
+
 class MDA(QWidget):
     """An example of using the MDAWidget to create and acquire a useq.MDASequence.
 
@@ -84,152 +90,177 @@ class MDA(QWidget):
 
         self.layout().addWidget(self.mda)
         self.layout().addWidget(live_viewer)
+class ConfigController(QWidget):
+    """AcquisitionEngine object for the napari-mesofield plugin.
+    This class is a subclass of the QWidget class.
+    The object connects to the Micro-Manager Core object instance and the napari viewer object.
 
+    _update_config: updates the experiment configuration from a new JSON file
 
-@magicgui(call_button='Start LED', mmc={'bind': pymmcore_plus.CMMCorePlus.instance()})   
-def load_arduino_led(mmc):
-    """ Load Arduino-Switch device with a sequence pattern and start the sequence """
-    
-    mmc.getPropertyObject('Arduino-Switch', 'State').loadSequence(['4', '4', '2', '2'])
-    mmc.getPropertyObject('Arduino-Switch', 'State').setValue(4) # seems essential to initiate serial communication
-    mmc.getPropertyObject('Arduino-Switch', 'State').startSequence()
+    run_sequence: runs the MDA sequence with the configuration parameters
 
-    print('Arduino loaded')
-
-@magicgui(call_button='Stop LED', mmc={'bind': pymmcore_plus.CMMCorePlus.instance()})
-def stop_led(mmc):
-    """ Stop the Arduino-Switch LED sequence """
-    
-    mmc.getPropertyObject('Arduino-Switch', 'State').stopSequence()
-
-class MainWidget(QMainWindow):
-    def __init__(self, core_object1: CMMCorePlus, core_object2: CMMCorePlus, cfg: ExperimentConfig):
+    launch_psychopy: launches the PsychoPy experiment as a subprocess with ExperimentConfig parameters
+    """
+    def __init__(self, mmc: pymmcore_plus.CMMCorePlus, cfg):
         super().__init__()
-        self.setWindowTitle("Main Widget with Two MDA Widgets")
-        self._meso_counter = count()
-        self._pupil_counter = count()
-        self._dhyana_metadata: dict[str, dict] = {}
-        self._thor_metadata: dict[str, dict] = {}
+        self._mmc = mmc
+        self.config: ExperimentConfig = cfg
 
-        # Create a central widget and set it as the central widget of the QMainWindow
-        central_widget = QWidget()
-        self.setCentralWidget(central_widget)
-        
-        # Set layout for the central widget
-        main_layout = QHBoxLayout(central_widget)
-        mda_layout = QVBoxLayout()
-        
-        # Create two instances of MDA widget
-        self.dhyana_gui = MDA(core_object1, cfg)
-        self.thor_gui = MDA(core_object2, cfg)
-        self.config = AcquisitionEngine(core_object1, cfg)
-        
-        # Add MDA widgets to the layout
-        mda_layout.addWidget(self.dhyana_gui)
-        mda_layout.addWidget(self.thor_gui)
-        main_layout.addLayout(mda_layout)
-        main_layout.addWidget(self.config)
-        
-        self.init_console()
-        toggle_console_action = self.menuBar().addAction("Toggle Console")
-        toggle_console_action.triggered.connect(self.toggle_console)
-        
-        self.dhyana_gui.mmc.mda.events.sequenceStarted.connect(self._dhyana_mda_start)
-        self.thor_gui.mmc.mda.events.sequenceStarted.connect(self._thor_mda_start)
-        self.dhyana_gui.mmc.mda.events.frameReady.connect(self._dhyana_save_frame_metadata)
-        self.thor_gui.mmc.mda.events.frameReady.connect(self._thor_save_frame_metadata)
-        self.dhyana_gui.mmc.mda.events.sequenceFinished.connect(self.save_meso_metadata)
-        self.thor_gui.mmc.mda.events.sequenceFinished.connect(self.save_pupil_metadata)
+        # Create main layout
+        self.layout = QVBoxLayout(self)
 
-    def _dhyana_mda_start(self) -> None:
-        """Called when the MDA sequence starts for Dhyana camera."""
-        self._meso_counter = count() # reset iterative counter
-        self._dhyana_metadata = {} # reset frame metadata storage
-        
-    def _thor_mda_start(self) -> None:
-        """Called when the MDA sequence starts for Dhyana camera."""
-        self._pupil_counter = count() # reset iterative counter
-        self._dhyana_metadata = {} # reset frame metadata storage
+        configUpdated = pyqtSignal(object)
 
-    def _dhyana_save_frame_metadata(self, image: np.ndarray, event: MDAEvent, frame_metadata: dict) -> None:
-        """Called each time a frame is acquired."""
-        frame_index = next(self._meso_counter)
-        self._dhyana_metadata[frame_index] = frame_metadata
-        
-    def _thor_save_frame_metadata(self, image: np.ndarray, event: MDAEvent, frame_metadata: dict) -> None:
-        """Called each time a frame is acquired."""
-        frame_index = next(self._pupil_counter)
-        self._thor_metadata[frame_index] = frame_metadata
+        # ==================================== GUI Widgets ===================================== #
 
-    def save_meso_metadata(self) -> None:
-        """Called when the MDA sequence ends for Dhyana camera."""
-        save_dir = Path('C:/dev/PyLab/tests')  # Make sure you use forward slashes or raw strings for Windows paths
-        save_dir.mkdir(parents=True, exist_ok=True)  # Ensure the directory exists
+        # 1. Selecting a save directory
+        self.directory_label = QLabel('Select Save Directory:')
+        self.directory_line_edit = QLineEdit()
+        self.directory_line_edit.setReadOnly(True)
+        self.directory_button = QPushButton('Browse')
 
-        # Construct the filename
-        time_str = datetime.datetime.now().strftime("%Y%m%d-%H%M%S")
-        filename = f'{time_str}_dhyana_frame_metadata.json'
+        dir_layout = QHBoxLayout()
+        dir_layout.addWidget(self.directory_label)
+        dir_layout.addWidget(self.directory_line_edit)
+        dir_layout.addWidget(self.directory_button)
 
-        # Save metadata to the json file
-        df = pd.DataFrame(self._dhyana_metadata)  # Transpose so that the rows are by time index
-        df = df.drop(index=['mda_event'])
-        df.to_json(save_dir / filename)
+        self.layout.addLayout(dir_layout)
 
-    def save_pupil_metadata(self) -> None:
-        """Called when the MDA sequence ends for Thor camera."""
-        save_dir = Path('C:/dev/PyLab/tests')  # Ensure consistency in save location
-        save_dir.mkdir(parents=True, exist_ok=True)  # Ensure the directory exists
+        # 2. Dropdown Widget for JSON configuration files
+        self.json_dropdown_label = QLabel('Select JSON Config:')
+        self.json_dropdown = QComboBox()
 
-        # Construct the filename
-        time_str = datetime.datetime.now().strftime("%Y%m%d-%H%M%S")
-        filename = f'{time_str}_thor_frame_metadata.json'
+        json_layout = QHBoxLayout()
+        json_layout.addWidget(self.json_dropdown_label)
+        json_layout.addWidget(self.json_dropdown)
 
-        # Save metadata to the json file
-        df = pd.DataFrame(self._thor_metadata)
-        df = df.drop(index=['mda_event']) # mda_event is type: MappingProxy, cannot be serialized to json
-        df.to_json(save_dir / filename)
-        
-    def _on_pause(self, state: bool) -> None:
-        """Called when the MDA is paused."""
+        self.layout.addLayout(json_layout)
 
-        
-    def toggle_console(self):
-        """Show or hide the IPython console."""
-        if self.console_widget and self.console_widget.isVisible():
-            self.console_widget.hide()
-        else:
-            if not self.console_widget:
-                self.init_console()
-            else:
-                self.console_widget.show()
-                
-    def init_console(self):
-        """Initialize the IPython console and embed it into the application."""
-        # Create an in-process kernel
-        self.kernel_manager = QtInProcessKernelManager()
-        self.kernel_manager.start_kernel()
-        self.kernel = self.kernel_manager.kernel
-        self.kernel.gui = 'qt'
+        # 3. Table widget to display the configuration parameters loaded from the JSON
+        self.layout.addWidget(QLabel('Experiment Config:'))
+        self.config_table = QTableWidget()
+        self.config_table.setEditTriggers(QTableWidget.AllEditTriggers)
+        self.config_table.horizontalHeader().setSectionResizeMode(QHeaderView.Stretch)
+        self.layout.addWidget(self.config_table)
 
-        # Create a kernel client and start channels
-        self.kernel_client = self.kernel_manager.client()
-        self.kernel_client.start_channels()
+        # 4. Record button to start the MDA sequence
+        self.record_button = QPushButton('Record')
+        self.layout.addWidget(self.record_button)
 
-        # Create the console widget
-        self.console_widget = RichJupyterWidget()
-        self.console_widget.kernel_manager = self.kernel_manager
-        self.console_widget.kernel_client = self.kernel_client
+        # ------------------------------------------------------------------------------------- #
 
-        # Expose variables to the console's namespace
-        self.kernel.shell.push({
-            'wdgt1': self.dhyana_gui,
-            'wdgt2': self.thor_gui,
-            'self': self
+        # ============ Callback connections between widget values and functions ================ #
 
-            # Optional, so you can use 'self' directly in the console
-        })
+        self.directory_button.clicked.connect(self._select_directory)
+        self.json_dropdown.currentIndexChanged.connect(self._update_config)
+        self.config_table.cellChanged.connect(self._on_table_edit)
+        self.record_button.clicked.connect(self.record)
 
-        
+        # ------------------------------------------------------------------------------------- #
+
+        # Initialize the config table
+        self._refresh_config_table()
+
+    # ============================== Private Class Methods ============================================ #
+
+    def _select_directory(self):
+        """Open a dialog to select a directory and update the GUI accordingly."""
+        directory = QFileDialog.getExistingDirectory(self, "Select Save Directory")
+        if directory:
+            self.directory_line_edit.setText(directory)
+            self._get_json_file_choices(directory)
+
+    def _get_json_file_choices(self, path):
+        """Return a list of JSON files in the current directory."""
+        import glob
+        self.config.save_dir = path
+        try:
+            json_files = glob.glob(os.path.join(path, "*.json"))
+            self.json_dropdown.clear()
+            self.json_dropdown.addItems(json_files)
+        except Exception as e:
+            print(f"Error getting JSON files from directory: {path}\n{e}")
+
+    def _update_config(self, index):
+        """Update the experiment configuration from a new JSON file."""
+        json_path_input = self.json_dropdown.currentText()
+
+        if json_path_input and os.path.isfile(json_path_input):
+            try:
+                self.config.load_parameters(json_path_input)
+                # Refresh the GUI table
+                self._refresh_config_table()
+            except Exception as e:
+                print(f"Trouble updating ExperimentConfig from AcquisitionEngine:\n{json_path_input}\nConfiguration not updated.")
+                print(e)
+
+    def _on_table_edit(self, row, column):
+        """Update the configuration parameters when the table is edited."""
+        try:
+            if self.config_table.item(row, 0) and self.config_table.item(row, 1):
+                key = self.config_table.item(row, 0).text()
+                value = self.config_table.item(row, 1).text()
+                self.config.update_parameter(key, value)
+        except Exception as e:
+            print(f"Error updating config from table: check AcquisitionEngine._on_table_edit()\n{e}")
+
+    def _refresh_config_table(self):
+        """Refresh the configuration table to reflect current parameters."""
+        df = self.config.dataframe
+        self.config_table.blockSignals(True)  # Prevent signals while updating the table
+        self.config_table.clear()
+        self.config_table.setRowCount(len(df))
+        self.config_table.setColumnCount(len(df.columns))
+        self.config_table.setHorizontalHeaderLabels(df.columns.tolist())
+
+        for i, row in df.iterrows():
+            for j, (col_name, value) in enumerate(row.items()):
+                item = QTableWidgetItem(str(value))
+                self.config_table.setItem(i, j, item)
+
+        self.config_table.blockSignals(False)  # Re-enable signals
+
+    # ----------------------------------------------------------------------------------------------- #
+
+    # ============================== Public Class Methods ============================================ #
+
+    def record(self):
+        """Run the MDA sequence with the global Config object parameters loaded from JSON."""
+
+        # Wait for spacebar press if start_on_trigger is True
+        wait_for_trigger = self.config.start_on_trigger
+        if wait_for_trigger:
+            print("Press spacebar to start recording...")
+            # self.launch_psychopy()
+            # Note: Implement key press detection suitable for PyQt5
+            # For example, using QEventLoop or custom dialog
+
+        # Run the MDA sequence
+        self._mmc.run_mda(
+            self.config.meso_sequence,
+            output=self.config.meso_data_path
+        )
+
+    def launch_psychopy(self):
+        """Launches a PsychoPy experiment as a subprocess with the current ExperimentConfig parameters."""
+        # Build the command arguments
+        args = [
+            "C:\\Program Files\\PsychoPy\\python.exe",
+            "F:\\jgronemeyer\\Gratings_vis_0.6.py",
+            f'{self.config.protocol}',
+            f'{self.config.subject}',
+            f'{self.config.session}',
+            f'{self.config.save_dir}',
+            f'{self.config.num_trials}'
+        ]
+
+        subprocess.Popen(args, start_new_session=True)
+    
+    def save_config(self):
+        """ Save the current configuration to a JSON file """
+        self.config.save_parameters()
+
+    #-----------------------------------------------------------------------------------------------#
 # if __name__ == "__main__":
 #     app = QApplication(sys.argv)
